refactor: 폴더명 변경 (register-keychain → keychain-register)
